refactor: split session pipeline tests from git insight scorer tests (#455)
